requirements.txt
Flask<1
paho-mqtt